completed settings dropdown and modal for delete account
